mapMenuWrapperとしてまとめた
